requirements.txt
# Flask==2.3.3
# Flask-SQLAlchemy==3.0.5
# Flask-Login==0.6.3
# Flask-WTF==1.1.1
# WTForms==3.0.1
# Werkzeug==2.3.7
# pywebview==4.4.1
# openpyxl==3.1.2
# reportlab==4.0.4
# Pillow==10.0.1

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1
reportlab==4.0.4
Pillow==10.0.1
numpy==1.24.3

